requirements.txt
git+https://github.com/adammotzel/glms.git
fastapi>=0.115.11
scikit-learn>=1.6.1
pydantic>=2.10.6
ipykernel>=6.29.5
uvicorn>=0.34.0
requests>=2.32.3
PyYAML>=6.0.2
python-dotenv>=1.0.1
pandas>=2.2.3
numpy>=2.2.3
